requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
keyboard==0.13.5
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
ninja==1.13.0
openai==1.99.9
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pynput==1.8.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
tifffile==2024.8.30
tiktoken==0.11.0
urllib3==2.5.0
youtube-transcript-api==1.2.2
gradio_client
moviepy==2.2.1
openai-whisper
pypinyin
